951_moves stfc seps export to own folder
