+++ apps/backend/src/factory/xlsx/SEPDataRow.ts
@@ -27,30 +27,6 @@ export function getDataRow(
   };
 }
 
-export function getStfcDataRow(
-  piName: string,
-  proposalAverageScore: number,
-  instrument: InstrumentWithAvailabilityTime,
-  sepMeetingDecision: SepMeetingDecision | null,
-  proposal: Proposal | null,
-  technicalReview: TechnicalReview | null,
-  sepProposal?: SEPProposal
-) {
-  return {
-    ...getDataRow(
-      piName,
-      proposalAverageScore,
-      instrument,
-      sepMeetingDecision,
-      proposal,
-      technicalReview,
-      sepProposal
-    ),
-    instrName: instrument.name,
-    feedback: sepMeetingDecision?.commentForUser,
-  };
-}
-
 export function populateRow(row: RowObj) {
   return [
     row.propShortCode ?? '<missing>',
@@ -62,21 +38,5 @@ export function populateRow(row: RowObj) {
     row.propReviewAvgScore ?? '<missing>',
     row.propSEPRankOrder ?? '<missing>',
     row.inAvailZone ?? '<missing>',
-  ];
-}
-
-export function populateStfcRow(row: RowObj) {
-  return [
-    row.propShortCode ?? '<missing>',
-    row.propTitle ?? '<missing>',
-    row.principalInv,
-    row.instrName ?? '<missing>',
-    row.instrAvailTime ?? '<missing>',
-    row.techReviewTimeAllocation ?? '<missing>',
-    row.sepTimeAllocation ?? row.techReviewTimeAllocation ?? '<missing>',
-    row.propReviewAvgScore ?? '<missing>',
-    row.propSEPRankOrder ?? '<missing>',
-    row.inAvailZone ?? '<missing>',
-    row.feedback ?? '<missing>',
   ];
 }
